requirements.txt
fastapi
uvicorn[standard]
greenlet
sqlalchemy
asyncpg
pgvector
google-generativeai
pydantic-settings
python-dotenv
sec-edgar-downloader
beautifulsoup4

